Fix issue with transaction date being a day earlier than entered date
In transaction-form.save(), when the new Transaction object is created a
date string of format YYYY-MM-DD is given. The Transaction constructor
would create a new Date object from this date string, which assumes
T00:00:00 for the time portion. Then, when a new date string was
created, a negative UTC offset would cause the date to be represented a
day earlier than what was originally entered.

The decision to have Transaction instead use a string instead of a Date
was because there isn't a way for the Angular Form to pipe a Date object
into the appropriate string format without considerable rework of the
transaction-form, and nothing else is making use of the Transaction
object, specificate the date property.